[BO] #5 JWT 토큰 기반 검증 및 식별키 추출 기능 구현
- 요청온 JWT 토큰에 대한 검증 작업 및 내부에 적재된 관리자 식별키 조회 기능 구현
- 관련 테스트 시나리오 추가
